refactor : 패키지 디렉터리 일관성있게 정리, 클래스 명 명시적으로 변경
